Feat [#7] DetailVC MVVM 리팩토링
시간별 collectionView 반영 필요
