+++ DO_SOPT_33_Assignment/DO_SOPT_33_Assignment/Presentation/Home/ViewController/HomeViewController.swift
@@ -158,7 +158,7 @@ extension HomeViewController: WeatherButtonDelegate {
         
         for index in 0..<homeViewModel.resultArray.count {
             let detailViewController = DetailViewController()
-            detailViewController.indexNumber = index
+            detailViewController.detailViewModel.indexNumber = index
             detailViewController.detailWeatherData = homeViewModel.resultArray[index]
             detailPageViewController.detailViewControllers.append(detailViewController)
         }
@@ -166,7 +166,7 @@ extension HomeViewController: WeatherButtonDelegate {
         let firstViewController = detailPageViewController.detailViewControllers[sender.indexNumber]
         detailPageViewController.pageViewController.setViewControllers([firstViewController], direction: .forward, animated: true, completion: nil)
         
-        detailPageViewController.detailViewControllers[sender.indexNumber].indexNumber = sender.indexNumber
+        detailPageViewController.detailViewControllers[sender.indexNumber].detailViewModel.indexNumber = sender.indexNumber
         
         navigationController?.pushViewController(detailPageViewController, animated: true)
     }
